# lib/api-zod/src/generated/types/generatePostInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TON News Bot API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratePostInputForceFormat } from './generatePostInputForceFormat';
import type { GeneratePostInputSourceType } from './generatePostInputSourceType';

export interface GeneratePostInput {
  topic?: string;
  sourceUrl?: string;
  sourceType?: GeneratePostInputSourceType;
  forceFormat?: GeneratePostInputForceFormat;
}