# lib/api-zod/src/generated/types/scheduleUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TON News Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduleUpdate {
  enabled?: boolean;
  intervalHours?: number;
  maxPostsPerDay?: number;
  autoPublish?: boolean;
}