# lib/api-zod/src/generated/types/sourceUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TON News Bot API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SourceUpdate {
  name?: string;
  url?: string;
  enabled?: boolean;
  isPrimary?: boolean;
  category?: string;
}